Added image uploading API

+++ src/main/java/com/codeup/rentlister/repositories/InquiriesRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface InquiriesRepository extends JpaRepository<Inquiries, Integer> {
 	Inquiries findInquiryById(int id);
-	List<Inquiries> findInquiriesByPropertyId(int id);
+	List<Inquiries> findInquiriesByPropertyId(long id);
 }
